requirements.txt
requests
beautifulsoup4
typer
openai
python-dotenv
